Solving Python Packging exercise
