+++ src/ahmedmd_diffusion_2d.egg-info/PKG-INFO
@@ -1,0 +1,5 @@
+Metadata-Version: 2.1
+Name: ahmedmd-diffusion-2d
+Version: 0.0.0
+Description-Content-Type: text/markdown
+License-File: LICENSE
