+++ src/ahmedmd_diffusion_2d.egg-info/SOURCES.txt
@@ -1,0 +1,10 @@
+LICENSE
+README.md
+setup.py
+src/ahmedmd_diffusion_2d/__init__.py
+src/ahmedmd_diffusion_2d/diffusion2d.py
+src/ahmedmd_diffusion_2d/output.py
+src/ahmedmd_diffusion_2d.egg-info/PKG-INFO
+src/ahmedmd_diffusion_2d.egg-info/SOURCES.txt
+src/ahmedmd_diffusion_2d.egg-info/dependency_links.txt
+src/ahmedmd_diffusion_2d.egg-info/top_level.txt
