+++ test.py
@@ -1,0 +1,2 @@
+from ahmedmd_diffusion_2d import diffusion2d
+diffusion2d.solve()
